everything server for REAME snippets
