+++ .pre-commit-config.yaml
@@ -36,3 +36,9 @@ repos:
         language: system
         files: ^(pyproject\.toml|uv\.lock)$
         pass_filenames: false
+      - id: readme-snippets
+        name: Check README snippets are up to date
+        entry: uv run scripts/update_readme_snippets.py --check
+        language: system
+        files: ^(README\.md|examples/.*\.py|scripts/update_readme_snippets\.py)$
+        pass_filenames: false
