+++ examples/servers/everything/pyproject.toml
@@ -1,0 +1,18 @@
+[project]
+name = "mcp-server-everything"
+version = "0.1.0"
+description = "Comprehensive MCP server demonstrating all protocol features"
+readme = "README.md"
+requires-python = ">=3.10"
+dependencies = [
+    "mcp",
+    "pydantic>=2.0",
+    "httpx",
+]
+
+[project.scripts]
+mcp-server-everything = "everything.__main__:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
