+++ examples/servers/everything/src/everything/__init__.py
@@ -1,0 +1,5 @@
+"""MCP Everything Server - Comprehensive example demonstrating all MCP features."""
+
+from .server import create_everything_server
+
+__all__ = ["create_everything_server"]
